Auto Request, when navigate between pages. #29
